chore: style the UI
